data: Add eos-icon-overrides.ini
This file specifies all the icon overrides that might be used
by the app center. We want the real apps to replace the fake
launchers when they are installed, but we can't do that through
a key in the appstream metadata because those apps might not
have been refined yet.

https://phabricator.endlessm.com/T18657