Include a --branch option to the scripts
The app helper installer assumes that an app's branch must be the
default one set for the remote we determine for the app. However, some
remotes do not have this default branch set, which results in the scrips
passing only the simple app IDs to GNOME Software, resulting in it
calling a search rather than showing the details page for the apps.

So this patch adds a branch option which will be used for calculating
the apps' unique ID, and thus open the apps' details page in GNOME
Software. If the branch is not passed, then it continues to look for
the remotes' default branch as before.

https://phabricator.endlessm.com/T21493