flatpak-autoinstall: migrate to EknServicesMultiplexer
The versioned EknServices apps have been replaced by one EknServicesMultiplexer
flatpak which contains the services for all currently supported SDKs which are
installed.

We must remove the old ones first so that the .service files are removed from
the Flatpak export path before the multiplexer is installed, as it exports the
same names.

https://phabricator.endlessm.com/T22063