flatpak-autoinstall: remove EndlessOS icon theme auto-install entry
Improve the robustness of upgrades from early 3.3 releases (where the
upgrader has gained support for Flatpak auto installs) to the latest
3.3 release. We found bugs with the automatic creation of Flatpak
remotes meaning that eos-sdk might not exist, which would prevent
the system upgrading past this point.

This auto-install entry is present in master so will be re-added when
the systems upgrade to 3.4.0, after which the users will have received
the fix ensuring that eos-sdk/flathub remotes are added successfully.

https://phabricator.endlessm.com/T22258