Add "Get Zoom" desktop launcher
Zoom is the dominant app for schools and universities forced to teach
remotely during the COVID-19 pandemic.

Unlike other launchers in this project, this does not add an appdata
file blacklisting the .desktop file. I couldn't find the Get Zoom
launcher in the app center when I tested this locally, so for whatever
unknown reason this seems no longer necessary. If wider testing confirms
this, we can drop the files for other launchers.

I also realised one can just use the absolute path to an icon, and so
rather than shipping a named 64×64 and 128×128 icon in the system path,
we can probably get away with linking directly to the cached icon from
Flathub, which is used when showing Zoom in the app center. This is a
bit imperfect because desktop icons are only 64×64 on low-dpi systems,
so the 128×128 icon (used here for the benefit of high-dpi systems) will
be downscaled. I couldn't see the difference on my low-dpi system, and
the simplification is pleasing.

https://phabricator.endlessm.com/T30089