flatpak-autoinstall: point entries for hack apps to flathub
Otherwise old system that upgrade will try to install the
Clubhouse from eos-apps and the hack apps will be removed
from eos-apps.

Also remove the update entries for the Clubhouse. Here are
the cases:

* User is on a version that predates the action to install
com.hack_computer.Clubhouse. Then we want them to automatically
install Clubhouse//stable from Flathub.

* User is on a version that post-dates the install action,
and has removed Clubhouse. We want no action to occur.

* User is on a version that post-dates the install action but
pre-dates an update action, and still has Clubhouse installed.
We would like them to update it. However, the installed
version will have branch eos3 and origin eos-apps. When we
remove Clubhouse from eos-apps it will not be possible to
update it, and this would cause the entire OS update to
fail. But we can't specify an update action from Flathub
because… the installed version is not from Flathub. So
in this case, all that we can do is, do nothing, and when
they boot into the new version the installed app will be
rewritten to get future updates from Flathub. Then we
wait for them to update it.

The install entry for the ProjectLibrary will be removed as well
and we only keep the uninstall entry.

https://phabricator.endlessm.com/T31133