Remove pre-eos4.0 autoinstalls
All systems which have upgraded from 3.x to 4.0 have been through the
checkpoint at 3.9.6, which contained all these actions, so in theory
they have all been applied.

It is *possible* that some actions failed to be applied after booting
into 3.9.6. However in practice this is unlikely since 3.9 also
introduced the eos-updater-flatpak-installer-fallback.{service,timer}
mechanism to retry any failed actions later in the boot process and then
every 24 hours.

Removing these actions has several positive consequences:

1. Copying an OS update to USB requires copying all apps that might be
   needed as part of upgrading to this OS version, because you never
   know what version you're upgrading from and what actions might have
   already been applied. This means that copying to USB always copies
   all of Hack, all of LibreOffice, Chromium, etc. at present just in
   case they are needed. This is not free in time or disk space. But
   these are *not* needed to update from 3.9.6 to ≥ 4.0.0; and if a
   system needs to be upgraded to 3.9.6, copying that OS version *will*
   copy those apps.

2. The image builder, by default, fails if there are any unapplied
   actions in the autoinstall files. This means that if you want to, for
   example, omit Hack in an image, you need to add an extra hook to undo
   the effect of the autoinstall file, or fake the stampfile.

We can't avoid the actions which actually are needed for 3.9.6 → 4.0.0+
not to be lossy (currently, adding Rhythmbox and Cheese), but we can
avoid the redundant ones now we have crossed a checkpoint.

https://phabricator.endlessm.com/T32818